==> code/01_OneAcreFund.Rmd <==
---
title: "Process OneAcreFund-DATA"
author: "Lena Bueschel"
date: "06/2022"
output: word_document
---


#EDITING ONACREFUND DATASET (AGRONOMIC SURVEY 2021)

## load packages
```{r}
library(tidyverse)
library(variables)
library(dplyr)
library(stringi)
library(stringr)
library(GGally)
library(scales)
library(pals)
library(rgdal)
library(raster)
library(viridis)
library(lubridate)

setwd("D:/LSG/Lena/IFPRI")
```


```{r}
data <- read.csv("D:/LSG/Max/IFPRI/CropCuts/1AF_direct_download/One_Acre_Fund_-_Agronomic_Survey_Data_-_2016_to_2021_rQqhQpy_data.csv", sep=";")
#insert path ONEACREFUND dataset

## only keep interesting columns
data <- select(data, unique_id, can_kg_ph, comp_kg_ph, country, crop, dap_kg_ph, field_latitude, field_longitude, 
                   harvest_date, hybrid, intercrop, lime_kg_ph, npk_kg_ph, plant_date, plot_hectares, season, 
                   seed_kg_ph, seed_type, urea_kg_ph, year, yield_kg_ph)

data <- filter(data, crop == 'maize')

## delete crop cuts without field coordinates, without yield, without harvest and plant date and which are not maize
data <- filter(data, field_latitude != 'NA', field_longitude != 'NA', yield_kg_ph != 0, 
                   harvest_date != "", plant_date != "", crop == "maize")

## delete one cropcut with erroneous coordinates
n <- which(data$field_longitude == max (data$field_longitude))
data <- data[-n,]

## convert dates
data$plant_date   <- as.Date(data$plant_date, format = "%d. %h %y")
data$harvest_date <- as.Date(data$harvest_date, format = "%d. %h %y")

## calculate growing season length
data$gs_length <- as.numeric(difftime(data$harvest_date, data$plant_date, units=c("days")))

## add survey_ID and sort by it
data$survey <- paste(data$country, data$year, sep="_")
data <- data[order(data$survey),]

## exclude implausible growing season lengths: less than 3 months, longer than one year
data <- data[which((data$gs_length > 89) & (data$gs_length < 365)),]

data_new <- NULL
## clean each survey (n=13) from yield and growing season length outliers
for (i in 1:length(unique(data$survey)))
{ data_sel <- data[which(data$survey==unique(data$survey)[i]),]

  quartiles <- quantile(data_sel$yield_kg_ph, probs=c(.25, .75), na.rm = FALSE)
  IQR <- IQR(data_sel$yield_kg_ph)
  Lower <- quartiles[1] - 1.5*IQR
  Upper <- quartiles[2] + 1.5*IQR 
  data_sel <- subset(data_sel, data_sel$yield_kg_ph > Lower & data_sel$yield_kg_ph < Upper)
  
  quartiles_gs <- quantile(data_sel$gs_length, probs=c(.25, .75), na.rm = FALSE)
  IQR_gs <- IQR(data_sel$gs_length)
  Lower_gs <- quartiles_gs[1] - 1.5*IQR_gs
  Upper_gs <- quartiles_gs[2] + 1.5*IQR_gs
  data_sel <- subset(data_sel, data_sel$gs_length > Lower_gs & data_sel$gs_length < Upper_gs)
  
  data_new <- rbind(data_new, data_sel)
}

## add ID
data_new$ID <- paste0("ID_", str_pad(1:NROW(data_new), 5, pad="0"))

## sort columns
data_new <- select(data_new, ID, country, year, survey, crop, field_latitude, field_longitude, 
               plant_date, harvest_date, gs_length, yield_kg_ph, plot_hectares, season, 
               can_kg_ph, comp_kg_ph, urea_kg_ph, lime_kg_ph, npk_kg_ph, dap_kg_ph, seed_kg_ph, seed_type, hybrid, intercrop)

## export
write.csv(data_new, "D:/LSG/Lena/IFPRI/OneAcreFund_DATASETS/CropCuts_1AF_finalselection.csv", row.names=F)

```

## how many crop cuts in total, and per country and year?
```{r}
dim(data_new)[1]
table(select(data_new, country))
table(select(data_new, year))
table(select(data_new, country, year))
```
## export

## growing season by country and year
```{r}
data_short <- read.csv("D:/LSG/Lena/IFPRI/OneAcreFund_DATASETS/final_selection_Neq_complete.csv")

data_short <- dplyr::select(data_short, ID, country, year, survey, gs_length, yield_kg_ph, plant_date, harvest_date, field_latitude, field_longitude)
data_short$DOY_plant <- yday(data_short$plant_date)
data_short$DOY_harvest <- yday(data_short$harvest_date)

summary <- NULL
for (s in 1:length(unique(data_short$survey)))
{ data_short_sel <- data_short[which(data_short$survey==unique(data_short$survey)[s]),]
  data_short_sel_summary <- data_short_sel %>% group_by(DOY_plant, DOY_harvest) %>% 
                              summarize(mean_yield=mean(yield_kg_ph), gs_length=mean(gs_length), number_cropcuts = length(unique(ID))) %>% as.data.frame()

plot1 <- ggplot(data_short_sel_summary, aes(x=DOY_plant, y=DOY_harvest, color=mean_yield, size=number_cropcuts)) + 
          geom_point(alpha=0.8) + 
          labs(x = "Plant Date (DOY)", y = "Harvest Date (DOY)") +
          ggtitle(unique(data_short$survey)[s]) +
          scale_color_gradientn(colours=cubicl(100)) +
          scale_size_continuous(range=c(2,6))

png(paste0("D:/LSG/Lena/IFPRI/plots_growseas/cluster_", unique(data_short$survey)[s], ".png"), height = 600, width = 800)
 print(plot1)
dev.off()
}
```
