==> code/03_Growing Season.Rmd <==
---
title: "Growing season new"
author: "Lena Bueschel"
date: "9 7 2022"
output: word_document
---

#PREPARATION 
library(tidyverse)
library(variables)
library(dplyr)
library(lubridate)
library(data.table)
library(zoo)

setwd("D:/LSG/Lena/IFPRI")


#CREATE CLIMATE DATA TABLES FOR EVERY YEAR WITH ALL CLIMATE VARIABLES PER DAY AND CROPCUT 
```{r}
#import timeseries 
data <- read.csv("D:/LSG/Lena/IFPRI/OneAcreFund_DATASETS/CropCuts_1AF_finalselection.csv")

for(y in 2016:2020){
  esa_soilmoist <- read.csv(paste0("D:/LSG/Lena/IFPRI/ESA_CCI_SM_v06.1/Export Survey2021/cc_maize/soilmoist_daily_",y,"_new.csv"))
  chirps <- read.csv(paste0("D:/LSG/Lena/IFPRI/CHIRPS p0.5/Export/precipitation_daily_",y,"_new.csv"))
  era_vswl <- read.csv(paste0("D:/LSG/Lena/IFPRI/ERA5_Land/Export VSWL1/vswl1_daily_",y,"_new.csv"))
  era_temp_mean <- read.csv(paste0("D:/LSG/Lena/IFPRI/ERA5_Land/Export SURFACE_TEMP/surface_temp_daily_mean_",y,"_new.csv"))
  era_temp_min <- read.csv(paste0("D:/LSG/Lena/IFPRI/ERA5_Land/Export SURFACE_TEMP/surface_temp_daily_min_",y,"_new.csv"))
  era_temp_max <- read.csv(paste0("D:/LSG/Lena/IFPRI/ERA5_Land/Export SURFACE_TEMP/surface_temp_daily_max_",y,"_new.csv"))
  era_prec <- read.csv(paste0("D:/LSG/Lena/IFPRI/ERA5_Land/Export TOTAL_PREC/tp_daily_",y,"_new.csv"))
  
  
  #get esa soilmoist and chirps into long format
  esa_soilmoist <- gather(data = esa_soilmoist,
                               key = "ID",
                               value = "esa_soilmoist",
                               2:length(esa_soilmoist),
                               factor_key = TRUE)
  
  chirps <- gather(data = chirps,
                        key = "ID", 
                        value = "chirps",
                        2:length(chirps),
                        factor_key = TRUE)
  
  #convert date and rename columns 
  esa_soilmoist$X <- as.character(esa_soilmoist$X)
  esa_soilmoist$X <- as.Date(esa_soilmoist$X, format = "%Y%m%d")
  names(esa_soilmoist)[1] <- "Date"
  
  chirps$X <- as.character(chirps$X)
  chirps$X <- as.Date(chirps$X, format = "%Y.%m.%d")
  names(chirps)[1] <- "Date"
  
  
  colnames(era_prec) <- c("ID", "Date", "era_prec")
  colnames(era_temp_max) <- c("ID", "Date", "era_temp_max")
  colnames(era_temp_mean) <- c("ID", "Date", "era_temp_mean")
  colnames(era_temp_min) <- c("ID", "Date", "era_temp_min")
  colnames(era_vswl) <- c("ID", "Date", "era_vswl")
  
  esa_soilmoist$Date <- as.character(esa_soilmoist$Date)
  chirps$Date <- as.character(chirps$Date)
  
  #join data 
  climate_data <- inner_join(era_prec, era_vswl, by = c("ID", "Date"))
  climate_data <- left_join(climate_data, era_temp_max, by = c("ID", "Date"))
  climate_data <- left_join(climate_data, era_temp_mean, by = c("ID", "Date"))
  climate_data <- left_join(climate_data, era_temp_min, by = c("ID", "Date"))
  climate_data <- left_join(climate_data, esa_soilmoist, by = c("ID", "Date"))
  climate_data <- left_join(climate_data, chirps, by = c("ID", "Date"))
  
  
  #get 1AF data 
  data_new <- filter(data, year == y)
  
  data_short <- dplyr::select(data_new, harvest_date, plant_date, ID, survey)
  
  #add to climate data and create DOY for dates 
  climate_data <- left_join(climate_data, data_short, by = "ID")
  
  climate_data$DOY_harvest <- lubridate::yday(climate_data$harvest_date)
  climate_data$DOY_plant <- lubridate::yday(climate_data$plant_date)
    
  print(which(is.na(climate_data$era_vswl)))
  
  #export 
  write.csv(climate_data, paste0("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_",y,"_new.csv"), row.names = F)
  
}

```

#BIND GROWING SEASON DATA TOGETHER
```{r}
#get data and combine
data_2016 <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_2016_new.csv")
data_2017 <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_2017_new.csv")
data_2018 <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_2018_new.csv")
data_2019 <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_2019_new.csv")
data_2020 <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_2020_new.csv")

head(data_2016)

data_complete <- rbind(data_2016, data_2017, data_2018, data_2019, data_2020)


#export 
write.csv(data_complete, "D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_complete.csv", row.names = F)

```

#GET GROWING SEASON AVERAGE 
```{r}
for(y in 2016:2020){
  
  climate_data <- read.csv(paste0("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_",y,"_new.csv"))

  #add year, plant_year and DOY 
  climate_data$plant_year <- substr(climate_data$plant_date, 1, 4)
  climate_data$year <- substr(climate_data$Date, 1, 4)
  climate_data$DOY <- lubridate::yday(climate_data$Date)
  
  #make one df per year 
  climate_data_1 <- filter(climate_data, year == y-1)
  climate_data_2 <- filter(climate_data, year == y)
  
  #add 365 days 
  climate_data_2$DOY <- climate_data_2$DOY +365
  
  #bind together 
  climate_data <- rbind(climate_data_1, climate_data_2)
    
  #make oen dataframe per year
  climate_data_prev <- filter(climate_data, plant_year == y-1)
  climate_data_new <- filter(climate_data, plant_year == y)
  
  #add 365 days - for unique DOYs
  climate_data_prev$DOY_harvest <- climate_data_prev$DOY_harvest + 365
  climate_data_new$DOY_plant <- climate_data_new$DOY_plant +365
  climate_data_new$DOY_harvest <- climate_data_new$DOY_harvest +365
  
  #bind together 
  climate_data <- rbind(climate_data_prev, climate_data_new)
  
  #setup results dataframe
  n <- length(unique(climate_data$ID))
  results <- as.data.frame(matrix(ncol = 7, nrow = n))
  colnames(results) <- c("esa_avg", "chirps_sum", "era_prec_avg", "era_temp_max_avg", "era_temp_mean_avg", "era_temp_min_avg", "era_vswl_avg")
  row.names(results) <- unique(climate_data$ID)
  

    for(j in 1:NROW(results)){

       climate_data_sel <- climate_data[which(climate_data$ID == row.names(results)[j]),]
  
        start <- unlist(climate_data_sel$DOY_plant[1])
        stop <- unlist(climate_data_sel$DOY_harvest[1])
        
          results$chirps_sum[j] <- sum(climate_data_sel$chirps[c(start:stop)], na.rm = TRUE)
          results$era_vswl_avg[j] <- mean(climate_data_sel$era_vswl[c(start:stop)], na.rm = TRUE)
          results$era_temp_mean_avg[j] <- mean(climate_data_sel$era_temp_mean[c(start:stop)], na.rm = TRUE)
          results$era_temp_max_avg[j] <- mean(climate_data_sel$era_temp_max[c(start:stop)], na.rm = TRUE)
          results$era_temp_min_avg[j] <- mean(climate_data_sel$era_temp_min[c(start:stop)], na.rm = TRUE)
          results$esa_avg[j] <- mean(climate_data_sel$esa_soilmoist[c(start:stop)], na.rm = TRUE)
    } 

#export dataframe 
write.csv(results, paste0("D:/LSG/Lena/IFPRI/GrowingSeasonData/growing_season_average_",y,"_new.csv"), row.names = F)

}

```
