==> code/04_KDD.Rmd <==
---
title: "KDD"
author: "Lena Bueschel"
date: "9 7 2022"
output: word_document
---

#1. HEAT WAVE FUNCTION
```{r}
DayHeatWave <- function(input, threshold, duration) {
  
  ## create internal list
  data_temp <- input

  # for each administrative unit
  for (o in 1:length(data_temp))  
  { data_temp[[o]]$DHW_threshold <- NA                    
    data_temp[[o]]$DHW_duration <- NA
    data_temp[[o]]$DHW_exceed <- NA                        
    data_temp[[o]]$DHW_count <- NA                         
    data_temp[[o]]$DHW_event <- NA                             
    data_temp[[o]]$DHW_result <- NA                           
  
    # for each day
    for (d in 1:NROW(data_temp[[o]]))  
    {   data_temp[[o]]$DHW_threshold[d] <- threshold
        data_temp[[o]]$DHW_duration[d]  <- duration
        ## check if TMAX is above heat threshold
        if((!is.na(data_temp[[o]]$era_temp_max[d])) & (data_temp[[o]]$era_temp_max[d] >= data_temp[[o]]$DHW_threshold[d])) { 
             ## record day heat event
             data_temp[[o]]$DHW_exceed[d]  <- 1                           
        
        if (d==1) {
        ## start counting wave days
        data_temp[[o]]$DHW_count[d]  <- 1
        } else { 
          if (!is.na(data_temp[[o]]$DHW_exceed[d]))
              data_temp[[o]]$DHW_count[d] <- sum(na.omit(c(data_temp[[o]]$DHW_count[d-1],1 ))) } } } 

      # for each day
      for (d in 1:NROW(data_temp[[o]]))
      { if ((!is.na(data_temp[[o]]$DHW_count[d]))&(!is.na(data_temp[[o]]$DHW_duration[d])))
        { if( (data_temp[[o]]$DHW_count[d] >= data_temp[[o]]$DHW_duration[d]) |
    (sum(na.omit( c(data_temp[[o]]$DHW_exceed[c((d+1):(d+(c(data_temp[[o]]$DHW_duration[d]-data_temp[[o]]$DHW_count[d]))))], 
                              data_temp[[o]]$DHW_count[d]) ))==data_temp[[o]]$DHW_duration[d]) ) { 
             data_temp[[o]]$DHW_event[d] <- 1
             data_temp[[o]]$DHW_result[d] <- data_temp[[o]]$era_temp_max[d]-data_temp[[o]]$DHW_threshold[d] 
           } } } }
return(data_temp)
}


test <- climate_data[(climate_data$ID == "ID_20759"),]

dim(test)
```

#2. get annual data and calculate kdd event & result for every cropcut for growing season
```{r}
library(data.table)
library(dplyr)
for(y in 2016:2020){ 
  y <- 2020
  climate_data <- as.data.frame(fread(paste0("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_",y,"_new.csv")))
  
  #delete cropcuts with NAs in the ERA columns
  delete <- unique(climate_data$ID[which(is.na(climate_data$era_temp_max))])
  climate_data <- climate_data[which(climate_data$ID %in% delete == FALSE),]

  #add year, plant_year and DOY 
  climate_data$plant_year <- substr(climate_data$plant_date, 1, 4)
  climate_data$year <- substr(climate_data$Date, 1, 4)
  climate_data$DOY <- lubridate::yday(climate_data$Date)
  
  #make one df per year 
  climate_data_1 <- filter(climate_data, year == y-1)
  climate_data_2 <- filter(climate_data, year == y)
  
  #add 365 days 
  climate_data_2$DOY <- climate_data_2$DOY +365
  
  #bind together 
  climate_data <- rbind(climate_data_1, climate_data_2)
    
  #make oen dataframe per year
  climate_data_prev <- filter(climate_data, plant_year == y-1)
  climate_data_new <- filter(climate_data, plant_year == y)
  
  #add 365 days - for unique DOYs
  climate_data_prev$DOY_harvest <- climate_data_prev$DOY_harvest + 365
  climate_data_new$DOY_plant <- climate_data_new$DOY_plant +365
  climate_data_new$DOY_harvest <- climate_data_new$DOY_harvest +365
  
  #bind together 
  climate_data <- rbind(climate_data_prev, climate_data_new)

  
  #create list -> each ID = 1 df for both years 
  data_list <- list()
  for(s in 1:length(unique(climate_data$ID))){
    data_list[[s]] <- climate_data[which(climate_data$ID == unique(climate_data$ID)[s]),]
  }
  
  #setup df for results 
  kdd <- as.data.frame(matrix(ncol = 1, nrow = length(data_list)))
  kdd$ID <- unique(climate_data$ID)

  #execute function
  results <- DayHeatWave(data_list, 29, 1)
  
  for(j in 1:NROW(results)){

    start <- unlist(results[[j]]$DOY_plant[1])
    stop <- unlist(results[[j]]$DOY_harvest[1])
          
    kdd$kdd_event[j] <- sum(results[[j]]$DHW_event[c(start:stop)], na.rm = TRUE)
    kdd$kdd_result[j] <- sum(results[[j]]$DHW_result[c(start:stop)], na.rm = TRUE)
  
  } 

  kdd$V1 <- NULL
  
  avg_data <- read.csv(paste0("D:/LSG/Lena/IFPRI/GrowingSeasonData/growing_season_average_",y,"_new.csv"))
  names(avg_data)[1] <- "ID"
  
  avg_data <- left_join(avg_data, kdd, by = "ID")
  
  write.csv(avg_data, paste0("D:/LSG/Lena/IFPRI/GrowingSeasonData/growing_season_average_",y,"_new.csv"), row.names = F)

}


```

#3. get daily kdd result for every cropcut
```{r}
library(data.table)
library(dplyr)


climate_data <- as.data.frame(fread(paste0("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_2020_new.csv")))
climate_data <- climate_data[,c(1,2,5)]


  #delete cropcuts with NAs in the ERA columns
  delete <- unique(climate_data$ID[which(is.na(climate_data$era_temp_max))])
  climate_data <- climate_data[which(climate_data$ID %in% delete == FALSE),]

  #create list -> each ID = 1 df for both years 
  data_list <- list()
  for(s in 1:length(unique(climate_data$ID))){
    data_list[[s]] <- climate_data[which(climate_data$ID == unique(climate_data$ID)[s]),]
  }



for(y in 2020){ 
  climate_data <- as.data.frame(fread(paste0("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_",y,"_new.csv")))
  
  #delete cropcuts with NAs in the ERA columns
  delete <- unique(climate_data$ID[which(is.na(climate_data$era_temp_max))])
  climate_data <- climate_data[which(climate_data$ID %in% delete == FALSE),]

  #create list -> each ID = 1 df for both years 
  data_list <- list()
  for(s in 1:length(unique(climate_data$ID))){
    data_list[[s]] <- climate_data[which(climate_data$ID == unique(climate_data$ID)[s]),]
  }

  #execute function
  results <- DayHeatWave(data_list, 29, 1)
  
  results_onetable <- NULL
  
  for(j in 1:NROW(results)){
    results_onetable <- rbind(results_onetable, results[[j]])
  } 
  
  write.csv(results_onetable, paste0("D:/LSG/Lena/IFPRI/GrowingSeasonData/climate_data_",y,"kdd_new.csv"))

}


```

#3. bind data together
```{r}
data_avg_2016 <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/growing_season_average_2016_new.csv")
data_avg_2017 <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/growing_season_average_2017_new.csv")
data_avg_2018 <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/growing_season_average_2018_new.csv")
data_avg_2019 <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/growing_season_average_2019_new.csv")
data_avg_2020 <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/growing_season_average_2020_new.csv")

data_avg_complete <- rbind(data_avg_2016, data_avg_2017, data_avg_2018, data_avg_2019, data_avg_2020)

write.csv(data_avg_complete, "D:/LSG/Lena/IFPRI/GrowingSeasonData/growing_season_average_complete.csv", row.names = F)

```
