==> code/06_df_setup.Rmd <==
---
title: "df_setup"
author: "Lena Bueschel"
date: "07/2022"
output: html_document
---

```{r}
oneaf <- read.csv("D:/LSG/Lena/IFPRI/OneAcreFund_DATASETS/CropCuts_1AF_finalselection.csv")
oneaf_N <- read.csv("D:/LSG/Lena/IFPRI/OneAcreFund_DATASETS/CropCuts_1AF_finalselection_Neq.csv")

growseas_avg <- read.csv("D:/LSG/Lena/IFPRI/GrowingSeasonData/growing_season_average_complete.csv")
```


```{r}
library(dplyr)
library(rgdal)
one_af_climate <- left_join(oneaf, growseas_avg, by = "ID")
one_afN_climate <- left_join(oneaf_N, growseas_avg, by = "ID")

one_af_climate <- select(one_af_climate, ID, survey, country, year, plant_date, harvest_date, gs_length, yield_kg_ph, hybrid, intercrop, esa_avg, 
                         chirps_sum,era_temp_mean_avg, era_temp_min_avg, era_temp_max_avg, kdd_event, kdd_result)

one_af_climate <- one_af_climate[complete.cases(one_af_climate),]


one_afN_climate <- select(one_afN_climate, ID, survey, country, year, field_latitude, field_longitude, plant_date, harvest_date, gs_length, 
                          yield_kg_ph, hybrid, intercrop, esa_avg, chirps_sum, era_temp_mean_avg, era_temp_min_avg, era_temp_max_avg, kdd_event, 
                          kdd_result, N_kg_ph)

one_afN_climate <- one_afN_climate[complete.cases(one_afN_climate),]

write.csv(one_af_climate, "D:/LSG/Lena/IFPRI/OneAcreFund_DATASETS/final_selection_complete.csv", row.names = F)
write.csv(one_afN_climate, "D:/LSG/Lena/IFPRI/OneAcreFund_DATASETS/final_selection_Neq_complete.csv", row.names = F)


shp <- SpatialPointsDataFrame(data=one_afN_climate, coords=one_afN_climate[,c(6,5)],
                                            proj4string = CRS("+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"))

writeOGR(shp, dsn = "D:/LSG/Lena/IFPRI/1AF_shapefiles", layer = "finalselection",
         driver = "ESRI Shapefile")


table(select(one_afN_climate, country, year))
```

#Z-Standardisierung
```{r}
one_afN_climate_stand <- one_afN_climate
one_afN_climate_stand$yield_kg_ph <- scale(one_afN_climate_stand$yield_kg_ph)
one_afN_climate_stand$N_kg_ph <- scale(one_afN_climate_stand$N_kg_ph)
one_afN_climate_stand$esa_avg <- scale(one_afN_climate_stand$esa_avg)
one_afN_climate_stand$chirps_sum <- scale(one_afN_climate_stand$chirps_sum)
one_afN_climate_stand$era_temp_mean_avg <- scale(one_afN_climate_stand$era_temp_mean_avg)
one_afN_climate_stand$era_temp_min_avg <- scale(one_afN_climate_stand$era_temp_min_avg)
one_afN_climate_stand$era_temp_max_avg <- scale(one_afN_climate_stand$era_temp_max_avg)
one_afN_climate_stand$kdd_event <- scale(one_afN_climate_stand$kdd_event)
one_afN_climate_stand$kdd_result <- scale(one_afN_climate_stand$kdd_result)
one_afN_climate_stand$gs_length <- scale(one_afN_climate_stand$gs_length)



write.csv(one_afN_climate_stand, "D:/LSG/Lena/IFPRI/OneAcreFund_DATASETS/final_selection_Neq_complete_zstand.csv", row.names = F)

```
